requirements.txt
Jinja2==3.1.2
wtforms==3.0.1
flask_wtf==1.0.1
requests==2.28.1
Flask==2.2.2
python-dotenv==0.21.0
google-auth==2.13.0
google-auth-oauthlib==0.7.0
oauthlib==3.2.2
pathlib==1.0.1
vendor==0.1
requests-oauthlib==1.3.1
Authlib==1.1.0
pip==22.3